% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oryzr.R
\name{foo}
\alias{foo}
\title{Test function.}
\usage{
foo(x)
}
\arguments{
\item{x}{Anything}
}
\value{
Returns the input. Bof sadfasdf  asdfkljsd dfjfkladjf sadfjsdkf
  asdfjsdf sjdfsdfjskldf fjjja asdlfjk asdfl
}
\description{
This function doesn't do anything.
}
\details{
This awesome function doesn't do anything, it just returns the input. Blahdy
blah blah blah.
}
\examples{
foo('abc')
foo(42)
foo(TRUE)
}

