% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicRangeToGeneName.R
\name{GenomicRangeToGeneName}
\alias{GenomicRangeToGeneName}
\title{GenomicRangeToGeneName}
\usage{
GenomicRangeToGeneName(seqnames, start, end)
}
\arguments{
\item{seqnames}{A \code{chr} containing the sequence (\emph{i.e.} 
chromosome) name. Must match the chromosome name in the MSUv7 GFF, 
\emph{e.g.} chromosome 1 == \code{'Chr1'}}

\item{start}{A \code{numeric} containing the starting coordinate on 
\code{seqnames}}

\item{end}{A \code{numeric} containing the end coordinat on \code{seqnames}}
}
\value{
Returns a \code{data.table} of annotated genes found in the MSUv7 
  GFF between \code{start} and \code{end} on chromosome \code{seqnames}.
}
\description{
Return information for all of the genes in a genomic region.
}
\seealso{
\code{LocToGeneName} for a description of the returned
  \code{data.table}.
}

