\name{oryzr-package}
\alias{oryzr-package}
\alias{oryzr}
\docType{package}
\title{
\packageTitle{oryzr}
}
\description{
\packageDescription{oryzr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{oryzr}
\packageIndices{oryzr}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{oryzr}

Maintainer: \packageMaintainer{oryzr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
