% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oryzr.R
\docType{package}
\name{oryzr}
\alias{oryzr}
\alias{oryzr-package}
\title{oryzr: Work with Oryza sativa gene identifiers.}
\description{
Convert between different types of gene identifiers for \emph{Oryza sativa} Nipponbare genes (MSU, Rap-DB, CSGNL, RefSeq mRNA).
}
\section{Functions}{

The oryzr functions ...
}

