% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SearchByGeneSymbol.R
\name{SearchByGeneSymbol}
\alias{SearchByGeneSymbol}
\title{SearchByGeneSymbol}
\usage{
SearchByGeneSymbol(GeneSymbols, shortLabels = FALSE,
  return.synonyms = FALSE)
}
\arguments{
\item{shortLabels}{Boolean (default \code{FALSE}). Tries to return a sensible
label \emph{e.g.} for plotting. See help for \code{LocToRefSeq} for more
information.}

\item{return.synonyms}{Boolean (default \code{FALSE}). Return a \strong{long}
  data.table including a column with the synonyms for each gene. See help for
  \code{LocToRefSeq} for more information.}

\item{\code{GeneSymbols}}{Character vector of gene symbols}
}
\value{
Returns a \code{data.frame} with MSU IDs as \code{rownames}, and
  columns RapID (RAP-DB gene identifier), symbols (CSGNL recommended symbols)
  and names (CSGNL recommended names), unless either of \code{shortLabels} or
  \code{return.synonyms} is \code{TRUE}. See help for \code{LocToRefSeq} for
  more information.
}
\description{
Search Rap-DB and oryzabase information by gene symbol and return identifiers
for potential matches.
}
\details{
Symbols are matched against the OryzabaseGeneListEn database, downloaded from
\href{http://www.shigen.nig.ac.jp/rice/oryzabase/download/gene}{Oryzabase}.
Rap-DB identifiers from the OryzabaseGeneListEn database are converted to MSU
identifiers using the table of RAP-DB and MSU IDs
\href{http://rapdb.dna.affrc.go.jp/download/irgsp1.html}{available from
Rap-DB}. \emph{n.b.} gene symbols are not necessarily unique so the
\code{data.table} returned by \code{SearchByGeneSymbol} must be manually
verified.
}
\examples{
SearchByGeneSymbol(c('FON1'))
SearchByGeneSymbol(c('SPL7'))
}

