% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{showDateRetrieved}
\alias{showDateRetrieved}
\title{showDateRetrieved}
\usage{
showDateRetrieved()
}
\value{
Returns a named character vector of date retrieved for each database.
}
\description{
Show the dates when the databases (RAP-DB to MSU IDs, OryzabaseGeneListEn and
Rice ID IRGSP Build5 database) were last downloaded to generate
\code{R/sysdata.rda}.
}
