% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oryzr.R
\docType{package}
\name{oryzr}
\alias{oryzr}
\alias{oryzr-package}
\title{oryzr: Work with Oryza sativa gene identifiers.}
\description{
Convert between different types of gene identifiers for \emph{Oryza sativa}
L. ssp. \emph{japonica} cv. Nipponbare genes (MSU, Rap-DB, CSGNL, RefSeq
mRNA).
}
\section{Functions}{
 \itemize{ \item \code{\link{LocToGeneName}}: Convert MSU
  identifiers to CSGNL gene names and symbols and generate a label for
  plotting if requested. \item \code{\link{LocToRefSeq}}: Convert MSU
  identifiers to Rap-DB identifiers and map them to RefSeq mRNA identifiers.
  \item \code{\link{SearchByGeneSymbol}}: Search Rap-DB and oryzabase
  information by gene symbol and return identifiers for potential matches.}
}

